refactored middlewares
