+++ middleware.js
@@ -1,7 +1,41 @@
+const ExpressError = require('./utility/ExpressError');
+const {campgroundSchema, reviewsSchema} = require('./schemas');
+const Campground = require('./models/campground');
 module.exports.isLoggedIn = (req, res, next) => {
     if(!req.isAuthenticated()){
         req.flash('error', 'You must be signed in')
         return res.redirect('/login')
+    }
+    next();
+}
+
+//client side validation for campground w/ Joi
+module.exports.validateCampground = (req, res, next) => {
+    const {error} = campgroundSchema.validate(req.body);
+    if(error){
+        const msg = error.details.map(el => el.message).join(',')
+        throw new ExpressError(msg, 400)
+    } else {
+        next();
+    }
+}
+
+module.exports.validateReview = (req, res, next) => {
+    const {error} = reviewsSchema.validate(req.body);
+    if(error){
+        const msg = error.details.map(el => el.message).join(',')
+        throw new ExpressError(msg, 400)
+    } else {
+        next();
+    }
+}
+
+module.exports.isAuthor = async(req, res, next) => {
+    const {id} = req.params;
+    const campground = await Campground.findById(id);
+    if(!campground.author.equals(req.user._id)){
+        req.flash('error', 'You do not have the permission to do that');
+        return res.redirect(`/campgrounds/${id}`);
     }
     next();
 }
