+++ routes/campgrounds.js
@@ -1,31 +1,10 @@
 const express = require("express");
 const router = express.Router();
 const catchAsync = require('../utility/catchAsync');
-const ExpressError = require('../utility/ExpressError');
 const Campground = require('../models/campground');
-const {campgroundSchema} = require('../schemas')
-const {isLoggedIn} = require('../middleware')
+const {isLoggedIn, validateCampground, isAuthor} = require('../middleware')
 
-//client side validation for campground w/ Joi
-const validateCampground = (req, res, next) => {
-    const {error} = campgroundSchema.validate(req.body);
-    if(error){
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg, 400)
-    } else {
-        next();
-    }
-}
 
-const isAuthor = async(req, res, next) => {
-    const {id} = req.params;
-    const campground = await Campground.findById(id);
-    if(!campground.author.equals(req.user._id)){
-        req.flash('error', 'You do not have the permission to do that');
-        return res.redirect(`/campgrounds/${id}`);
-    }
-    next();
-}
 
 
 router.get('/', catchAsync(async (req, res) => {
